bp/requirements.txt
termcolor
virtualenv
requests
pyOpenSSL
ndg-httpsclient
pyasn1
feedparser
bs4
stanfordcorenlp
pony
pymysql
celery
redis==2.10.5
flower
docker
jinja2==2.8
pyyaml
cryptography
google-cloud-language
gensim
numpy
sklearn
word2vec
